split GeomtryUtils in two files
