+++ alchemist/alchemist-implementationbase/src/main/kotlin/it/unibo/alchemist/model/implementations/geometry/GeometryUtils.kt
@@ -2,15 +2,11 @@
 package it.unibo.alchemist.model.implementations.geometry
 
 import it.unibo.alchemist.model.implementations.positions.Euclidean2DPosition
-import it.unibo.alchemist.model.interfaces.geometry.Vector
 import it.unibo.alchemist.model.interfaces.geometry.euclidean.twod.Euclidean2DSegment
 import org.danilopianini.lang.MathUtils.fuzzyEquals
 import java.awt.Shape
 import java.awt.geom.PathIterator
 import kotlin.math.atan2
-import kotlin.math.sqrt
-import kotlin.math.pow
-import kotlin.math.acos
 
 /**
  * Computes the angle with atan2(y, x).
@@ -45,42 +41,11 @@ fun Shape.vertices(): List<Euclidean2DPosition> {
 }
 
 /**
- * Multiplies each coordinate of a vector for a scalar number n.
- */
-fun Euclidean2DPosition.times(n: Double) = Euclidean2DPosition(x * n, y * n)
-
-/**
- * Normalizes the vector.
- */
-fun Euclidean2DPosition.normalize(): Euclidean2DPosition = times(1.0 / sqrt(x * x + y * y))
-
-/**
- * Resizes the vector in order for it to have a length equal
- * to the specified parameter. Its direction and verse are preserved.
- */
-fun Euclidean2DPosition.resize(newLen: Double): Euclidean2DPosition = normalize().times(newLen)
-
-/**
- * Find the normal of a vector.
- */
-fun Euclidean2DPosition.normal(): Euclidean2DPosition = Euclidean2DPosition(-y, x)
-
-/**
- * Computes the z component of the cross product of the given vectors.
- */
-fun zCross(v1: Euclidean2DPosition, v2: Euclidean2DPosition) = v1.x * v2.y - v1.y * v2.x
-
-/**
- * Dot product between bidimensional vectors.
- */
-fun Euclidean2DPosition.dot(v: Euclidean2DPosition) = x * v.x + y * v.y
-
-/**
  * Checks whether the given edge is inside a rectangular region described by an origin
  * point and width and height values (only positive).
  */
-fun isInBoundaries(e: Euclidean2DSegment, origin: Euclidean2DPosition, width: Double, height: Double) =
-    isInBoundaries(e.first, origin, width, height) && isInBoundaries(e.second, origin, width, height)
+fun isInBoundaries(edge: Euclidean2DSegment, origin: Euclidean2DPosition, width: Double, height: Double) =
+    isInBoundaries(edge.first, origin, width, height) && isInBoundaries(edge.second, origin, width, height)
 
 /**
  * Checks whether the given point is inside a rectangular region described by an origin
@@ -130,34 +95,6 @@ fun areCollinear(p1: Euclidean2DPosition, p2: Euclidean2DPosition, p3: Euclidean
         fuzzyEquals((m * p3.x + q), p3.y)
     }
 }
-
-/**
- * Finds the magnitude of a vector.
- */
-fun <V : Vector<V>> Vector<V>.magnitude(): Double {
-    var sum = 0.0
-    for (d in 0 until dimensions) {
-        sum += getCoordinate(d).pow(2)
-    }
-    return sqrt(sum)
-}
-
-/**
- * Computes the dot product between two vectors.
- */
-fun <V : Vector<V>> Vector<V>.dot(other: V): Double {
-    var dot = 0.0
-    for (d in 0 until dimensions) {
-        dot += getCoordinate(d) * other.getCoordinate(d)
-    }
-    return dot
-}
-
-/**
- * Computes the angle in radians between two vectors.
- */
-fun <V : Vector<V>> Vector<V>.angleBetween(other: V): Double =
-    acos(dot(other) / (magnitude() * other.magnitude()))
 
 /**
  * Checks whether the segment is aligned to the x axis.
